Dressrosa/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
httpx==0.28.1
pytest==8.4.2
sqlalchemy==2.0.43
alembic==1.16.5
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
